The hash code is now taken from the last 32 bits of the hash.
The last 32 bits are selected because the first bits might include zero
padding when the hash length is not a multiple of 8 bits.

All current algorithms have hash lengths that are a multiple of 8 bits,
so there is no need to perform a release now.  This is just to future-
proof the implementation.