Created SQL function "com.aoindustries.security"."Identifier.validate".
Although it is very simple in that it merely ensures both values are
non-null and could be accomplished in per-table check "IS NOT NULL"
check constraints, this follows the pattern of other composite types.
Composite types are responsible for their own validation, and will also
become DOMAIN in PostgreSQL 11+, further solidifying this.